index xml source from harvest object or xml location url
index xml source when available. first tries to index harvested object as this is already in the database. if harvest object is not xml (came from another ckan and thus is json) or just not available then it tries to read the xml file from the xml location url field.

indexed xml is stored in extras_harvest_document_content which only exists in the solr index and is also included in the 'All Text' search. To see the contents of this field you must query it with the 'fl' attribute set.

eg. https://localhost/api/3/action/package_search?fl=extras_harvest_document_content