rework the css
merging in master messed up this PR so I re did it. I made some modifications to the original styling as I think the current pacific wp site css is currently quite bad. css is a merge of Pacific, Atlantic, and national theming.

I added a no id to the language picker in the top header to more easily allow for styling of this item differently to other items in the header.

Cleaned up the css and removed several redundant, duplicated,  and unused rules